splitting up circuit
