+++ qiskit/circuit/library/__init__.py
@@ -104,12 +104,12 @@ Quantum Complexity Advantage Circuits
 
 from .blueprintcircuit import BlueprintCircuit
 from .generalized_gates import Diagonal
+from .generalized_gates import Permutation
 from .boolean_logic import (
-    Permutation,
+    AND,
+    OR,
     XOR,
     InnerProduct,
-    OR,
-    AND,
 )
 from .basis_change import QFT
 from .arithmetic import (
