+++ qiskit/circuit/library/generalized_gates/__init__.py
@@ -12,6 +12,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""The basis change circuits."""
+"""The generalized gates."""
 
 from .diagonal import Diagonal
+from .permutation import Permutation
